moving inttobytes3 to util
