+++ artemis/src/main/java/tech/pegasys/artemis/state/BeaconState.java
@@ -23,7 +23,6 @@ import tech.pegasys.artemis.datastructures.BeaconChainState.ShardCommittee;
 import tech.pegasys.artemis.datastructures.BeaconChainState.ShardReassignmentRecord;
 import tech.pegasys.artemis.datastructures.BeaconChainState.ValidatorRecord;
 import tech.pegasys.artemis.ethereum.core.Hash;
-import tech.pegasys.artemis.util.bytes.Bytes3;
 import tech.pegasys.artemis.util.uint.UInt64;
 
 import java.util.Arrays;
@@ -70,22 +69,6 @@ public class BeaconState {
   static class BeaconStateHelperFunctions {
 
 
-    /**
-     * Converts int to Bytes3.
-     *
-     * @param seed  converted
-     * @return      converted Bytes3
-     * @throws IllegalArgumentException if seed is a negative value.
-     */
-    @VisibleForTesting
-    static Bytes3 intToBytes3(int seed) {
-      checkArgument(seed > 0, "Expected positive seed but got %s", seed);
-      byte[] bytes = new byte[3];
-      bytes[0] = (byte) (seed >> 16);
-      bytes[1] = (byte) (seed >> 8);
-      bytes[2] = (byte) seed;
-      return Bytes3.wrap(bytes);
-    }
 
     /**
      * Converts byte[] to int.
