+++ util/src/main/java/tech/pegasys/artemis/util/bytes/Bytes3.java
@@ -33,6 +33,22 @@ public interface Bytes3 extends BytesValue {
   Bytes3 ZERO = wrap(new byte[3]);
 
   /**
+   * Converts int to Bytes3.
+   *
+   * @param seed  converted
+   * @return      converted Bytes3
+   * @throws IllegalArgumentException if seed is a negative value.
+   */
+  static Bytes3 intToBytes3(int seed) {
+    checkArgument(seed > 0, "Expected positive seed but got %s", seed);
+    byte[] bytes = new byte[3];
+    bytes[0] = (byte) (seed >> 16);
+    bytes[1] = (byte) (seed >> 8);
+    bytes[2] = (byte) seed;
+    return Bytes3.wrap(bytes);
+  }
+
+  /**
    * Wraps the provided byte array, which must be of length 3, as a {@link Bytes3}.
    *
    * <p>
